requirements.txt
fastapi>=0.111.0
uvicorn>=0.28.0
httpx>=0.27.0,<0.28
supabase>=2.24.0
openai>=1.40.0
langgraph>=1.0.3
apify-client>=2.3.0
python-dotenv>=1.0.1
pandas>=2.2.2
googlemaps>=4.10.0
reportlab>=4.0.8
fastmcp>=0.0.7
pytest>=7.4
